#56 💷  refactor: utils 폴더 생성 및 주석 제거
