+++ src/atoms/atom.tsx
@@ -66,11 +66,6 @@ export const userUidAssetData = atom<UserAssetData>({
   default: {},
 });
 
-export const coinCurrentPrice = atom({
-  key: "coinCurrentPrice",
-  default: 0,
-});
-
 export const pageAtom = atom<number>({
   key: "pageAtom",
   default: 1,
